Check if device is network or serial
